Getting top 10 user video links
Adding function for getting users video links from top.

I wasnt able to get link from the same http request as the rest of informations, cos the part with videos is loaded afterwards.

I had to find another http request which respond is Json file, where are included video links.

But the script depends on the link with Json request and isnt possible find this link by itself, because the link contains GET parameters, such as _signature, which I am not able to find out how is it going to change for e.g. different user or more videolinks.

BASICALLY it gets top 10 video links, but the link for the json must be find manually